CUDA: Move ufunc classes from dispatcher to vectorizers
